Fixed Cost function in QP formulation generation.
The cost function adds cost on the inital state x_0 that is constraint to be equal to x_init.
Now we just make the initial cost to be 0. This will result in a sparser system and therefor in faster execution and less memory ussage. Even though the argument can be made that the change is negligible.

I hope this change helps anybody that uses this code to generate a QP problem from an MPC one.